tested/refactored - test again before submit / doesn't support single quotes
