+++ filesActions/readFileToConsole.js
@@ -1,12 +1,9 @@
 import { createReadStream } from "node:fs";
-import { join } from "node:path";
+import { createCorrectPath } from "../utils/createCorrectPath.js";
 import { isAccessible } from "../utils/isAccessible.js";
 
 export const readFileToConsole = async (currentDir, cmd) => {
   let targetPath = cmd.substr(4);
-  // let newPath = join(currentDir, targetPath);
-  // if (targetPath.includes(":")) newPath = targetPath;
-  // if (targetPath[0] === "/" || targetPath[0] === "\\") newPath = targetPath;
   const newPath = createCorrectPath(currentDir, targetPath);
   if (!(await isAccessible(newPath))) {
     console.log("Invalid input");
